Improvements to the submission flow tests.
Getting hold of the application context via grails.util.Holders in
subclasses of FlowBase, which extends WebFlowTestCase results in
IllegalAccessErrors being thrown.

Mitigate against this behaviour by getting grailsApplication in the
static initialiser of the test class and then passing grailsApp.mainCtx
as an argument to the constructor of each FlowBase subclass.

End of black magic.