Enable dbCreate in production.
This enables Jummp to cope with the scenario of running in production
against a fresh database. Without this setting, Jummp expects manual
creation of the database structure.

The potential issue with this approach is that now Hibernate will try to
automagically update the database schema so that it matches our domain
model but, since it takes a very conservative, drop-nothing, approach it
will fail and leave the database in an inconsistent state should there
be any changes that impact referential integrity.

Therefore it is imperative that prior to any deployment that does not
expect a clean database we ensure that all outstanding migrations are ran.