JUM-17 Explicitly added perf4j as a dependency.
The perf4j plugin comes with a jar in charge of log4j logging in its lib
folder. The jar is not picked up if the Grails wrapper is used.

All our plugins depend on the security plugin, so if the latter contains
the perf4j dependency, all our other plugins, and in turn, the main app,
will obtain it through transitivity.