JUM-18 use GORM-aware tasks inside search service.
That way, we avoid having to manually bind a new Hibernate session to each
worker thread created by Grails promises.

Also simplified the way in which the list of revisions is obtained
during the regeneration of the index: since the action can only be
performed by admins, we can bypass the security overhead in modelService
with a direct call to the Revision domain object.