JUM-38 make vcsIdentifier contain the name of the container.
If we don't store the name of the container in the database, then
it's not possible to know in which container to look for the model
folder.

Also expanded the logic for working out the name of the model folder
container so that it correctly returns that 'aba' follows from 'aaz'
and that it recognises that the folder 'aaaa' was created after 'zzz'.

Finally, turned FileSystemService tests into unit tests it better
resembles the way in which they are written.