Fixed confusing log message.
If the audit entry's success status has been updated in memory as a
result of business logic, then it does make sense for audit.isDirty() to
return true, because it's different from what's in the database.

Keep calm and carry on saving to the database the updated entry rather
than issuing a warning.