Refactored mechanism for creating repository files.
Previously we would validate that the repository files existed in the
validation constraints for RepositoryFile. The problem is that the
validation is performed every time the model or the revision is saved,
including situations like archiving models, adding publication
information or publishing  them.

A further problem is that, as we don't store the full path of the
repositoryFile in the database, model validation fails because the
repositoryFile's path does not point to something that exists.

This commit therefore changes the logic so that the file path and mime
type of a repository file are only checked/enforced when a new model or
a new revision is created, as repository file information is not updated
anywhere.

It also makes the code for generating repositoryFiles from the
corresponding command objects more DRY.