JUM-170 save models/revisions in a dedicated transaction.
The OpenSessionInViewInterceptor fires a transaction at the start of
each request and commits it once the controller deals with it. This
means that for the duration of the request, unless otherwise specified,
new model submissions (and model updates) are only available within the
Hibernate session corresponding to that request, assuming a sane
transaction isolation level like READ_COMMITTED or higher.

Because the indexing process is now synchronous, we need to ensure that
the new model/revision is in the database before the request completes.
To achieve this we use a dedicated transaction with propagation
REQUIRES_NEW for saving the new model or revision. The effect is that we
ensure that the new content is saved atomically, and not rolled back as
a result of any other failures, which is also beneficial.