Call searchService directly after submission with detached objects.
Due to a bug in Spring's application event multicaster, the
model/revision created event is sent twice to the listener bean
(indexingListener). This does not happen when the application starts, so
it's difficult to mitigate against.

Evaluating the applicationEventMulticaster bean, one can see that its
registry has duplicate definitions of other listener beans, like bives'
revisionCreatedListener or the postLoggingAdvice which intercepts public
method calls annotated with a specific annotation.

Work around this by not relying on Spring's event bus at the end of the
submission flow and using searchService directly from modelService.

Also, previously we tried to get doAddValidatedRevision and
doUploadValidatedModel to execute in a different transaction but in the
same Hibernate session as the current request. This was achieved by
fetching the modelService proxy and calling the respective methods and
the rationale for using the proxy was to enforce our transactional
settings -- PROPAGATION_REQUIRES_NEW. As the Spring AOP reference
explains, the AOP interceptors do not kick in calls between methods
belonging to the same class! We've now simplified this by simply passing
a TransactionDefinition to the withTransaction block. Sadly, there is no
change in behaviour: we're still seeing a different Hibernate session
for the new transaction, which means that the objects returned by
doAddValidatedRevision and doUploadValidatedModel are detached from the
session.

Finally, I wrapped the code in SearchService.indexRevision within a
withSession block to mitigate the problem above whereby the detached
objects' associations are not necessarily initialised.