Don't use DetachedCriteria.list() when expecting 1 result.
When DetachedCriteria queries return a list of results, one actually
gets a list of lazy proxies that are not automatically unwrapped.

This causes issues when objects are used outside of a session.
Therefore, the list() method call was changed to get(), as in the case
of a single result, GORM unwraps the proxy for us.