JUM-193 use READ_COMMITTED tx isolation level for model update.
Due to the fact that modelService.addValidatedRevision, called both
during model update and as a result of the model being annotated, fires
nested transactions, we need the first one to be able to read data that
has been modified by the second one. This does not happen if the
isolation level is REPEATABLE_READS, which is the default in MySQL.

This isolation level needs to be specified for the service method that
is first called from within a controller, because that's when the
transaction context will be created by the transaction interceptor. All
service methods invoked downtream usually have transaction propagation
REQUIRED, meaning they will actually reuse the original transaction.