Prefer Grails @Transactional in submission/metadata services.
This annotation results in AST transformations wrapping target methods
or classes into GrailsTransactionTemplate callbacks. This means
that the transactional behaviour is embedded in the binaries and that
we're no longer relying on Spring AOP to process services and generate
loads of dynamic proxies at application startup. Conveniently, there is
also a Grails @NotTransactional method-level annotation that can be
employed to disable this behaviour.

This commit standardises the way in which we trigger the transactional
behaviour across all services dealing with new content, whether as a
result of the submission flow or the annotation editor.