Revisited transactional behaviour for modelHistoryService.
Use a dedicated transaction for editing the user's history. Exceptions
from this component should not impact other units of work. Because of
the dedicated transactional context, we no longer need to force flushing
in order for the changes to be reflected in the database because the end
of the transaction will trigger a commit.

Also simplified the queries we perform when retrieving or modifying the
user's history.