Worked around static compilation issue with getProperties.
Any Groovy object has a dynamically-generated .getProperties() method
which dumps all the instance variables of the object into a map.

In static compilation contexts it seems calling getProperties on an
object that has an instance variable named properties results in the
invocation of the dynamically-generated method as opposed to the
accessor.

Used Java reflection to invoke the correct method in order to obtain the
properties associated with an annotation section.