fixed deletion of stale data in annotation editor.
If the annotations present in the current version of the model cannot be
rendered by the editor, they should be removed from memory so as to
avoid carrying them into the new version of the model without the user
realising that this is the case.

This deals with the situation when the information coming from the
annotation input source changes.