Updating the descriptions of the additional files
We use an input element to store the additional file name instead
of displaying it in plain text only. That helps us collecting the
file name and refering to it when editing the descriptions. Because
the additional files are displayed in a table, we do not know which
file is in change. Tracking of the value of these input elements enables
us to follow the corresponding descriptions.

To edit the description, an input element is inserted between the
file name and discard button like adding an additional file. It was
decorated by increasing the width so that its width fits the cell
width.

In order to make these input elements programmable, it should use
a counter to number their indices.