Make a lot of modifications on file upload page
- Propose a map maintaining the fresh list of additional files. It
will be updated when adding, discarding or updating the description
is involked.
- Initialise this map at loading page.
- Retrieve the file name at clicking on Discard link, then delete
the element which key is identical with the file name.
- Set value attribute of the input file element to the file name.
- Realise the change event on the input text of description: get
the file name set to the input file element on the first column of
the current row, update the value of the element which key is the
file name by the newly-typed description, do not forget to populate
date to the referenced tags.