Deal with modifying the description of additional file
On the one hand, we get the map of additional files from the client.
This map contains all working additional files. We do not care what
have just added or are already existing.

On the other hand, we transfer the map of <String, String> to the one
of <File, String> with the purpose of sending them to SubmissionService.

Finally, we update the additional_files variable in workingMemory
and the list of deleted file names as well.