Update the repository files
The repository_files variable stores the working files. In
handleSubmission method, we expect to get the list of RFTC that will be
added and the list of file names which will be deleted. Then they are
passed to storeRFTC to store in workingMemory.

Change the additional files.

Comment at some source code lines in order to make them clearer.