Fix the error of parsing string
To help controller collect data from view via params object, we are using
an input element to store exchangeable data as text/string. In the case
of uploading files, additional files and associated descriptions are
maintained in an array that converted to a string, assigned it to the
input element later. To avoid using string-based operations with this
string because clients could type whatever they want, we use JSON
functions to deal with the data rather pure text/string. Now the data
sending to controller is a JSON string. In controller, the data are parsed
by JsonSlurper module.