Use LibPharmML to update the PharmML model description.
Our old approach, which used a custom XML parser didn't handle
correctly the situation when a model didn't have a description. It would
try to insert a Description element before the Name element, which is
invalid according to the PharmML spec.

Using LibPharmML's setDescription() method means that we don't need to
worry about any of this.