Using selected publication link provider to display rather the publication existing in the updating model
We store a ModelTC in workingMemory as the updating model. Behind every step of adding publication,
its publication is updated. That is the reason why we cannot get back the information of the previously
selected publications (i.e. when you want to change to mind). Moreover, the updating authors will be
removed out and overrided if you switch to other publication provider. If the existing publication
could not been shown in the publication editor form, submitter doesn't know who are authors of this
publication and he/she will add more authors who inherently existing into the database. We display the
currently selected publication object getting from the publication map intead of.