Improve the publication adding flow
We proposed a class, name PublicationDetailExtractionContext, which is used for:
1. checking whether the publication is existing or not. If it is already in the database, a warning message should be shown
aiming at telling users pay attention to their modifications. Because the publication is stored in the database, it is potentially
linked to other models. These updates will be influenced on the other models.
2. keeping the publication in working memory. We used PublicationTransportCommand stored as in-working publication objects in the
publication map. Because we would have the feature (1), PublicationTransportCommand objects are now replaced by PublicationDetailExtractionContext
ones.